requirements/base.txt
Django>=2
django-appconf<=1.0.3
django-model-utils>=3